moved common UI to core module
