+++ app/src/main/java/com/example/recipebook/screens/CategoriesScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.Text
@@ -23,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.recipebook.DishCategory
 import com.example.recipebook.R
-import com.example.recipebook.screens.modules.CategoryCard
+import com.example.core.ui.CategoryCard
 
 @Preview(showBackground = true)
 @Composable
@@ -54,7 +53,7 @@ fun CategoriesScreen() {
                     .fillMaxSize()
             ) {
                 items(categoryList) { category ->
-                    CategoryCard(category.name)
+                    com.example.core.ui.CategoryCard(category.name)
                 }
             }
         }
