+++ core/src/main/java/com/example/core/ui/BottomBar.kt
@@ -1,4 +1,4 @@
-package com.example.recipebook.screens.modules
+package com.example.core.ui
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Done
@@ -6,9 +6,7 @@ import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.recipebook.R
 
 @Preview(showBackground = true)
 @Composable
